requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
pyngrok>=7.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.121.0
python-dotenv>=1.0.0
google-search-results==2.4.2
